requirements.txt
opencv-python==4.12.0.88
PyTurboJPEG==1.8.2
linuxpy==0.23.0
numpy<2.3.0
onnxruntime-gpu==1.23.2
